requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==22.1
django-webpack-loader==1.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
gunicorn==20.1.0
pillow==11.2.1
psycopg2==2.9.3
PyJWT==2.9.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
